added some code

+++ src/pages/Courses.js
@@ -3,30 +3,44 @@ import React, { useState, useEffect } from 'react';
 import Header from '../components/Header';
 // import Button from '../components/Button';
 // import AddDataForm from '../components/AddDataForm';
+import { toast } from 'react-toastify';
+import 'react-toastify/dist/ReactToastify.css';
 import DataTable from '../components/DataTable';
 import { FaEdit, FaTrash } from 'react-icons/fa';
 import { Button, Table } from 'react-bootstrap';
 import AddDataForm from '../components/AddDataForm';
-import { fetchData } from '../App'; // Import the fetchData utility function
 
 const Courses = () => {
   const [courseData, setCourseData] = useState([]);
   const [showAddForm, setShowAddForm] = useState(false);
   const [editData, setEditData] = useState(null);
   const [deleted, setDeleted] = useState(false);
+
+  // to fetch the data
+  const getData = async () => {
+    try {
+      const response = await fetch('https://asxucwg1u7.execute-api.us-east-1.amazonaws.com/dev-test-1/dspfetchalldata')
+      const jsonData = await response.json()
+      const actualData = JSON.parse(jsonData.body);
+      setCourseData(actualData);
+      console.log(actualData)
+    } catch (error) {
+      // Handle error
+    }
+  };
   
   useEffect(() => {
-    const getData = async () => {
-      try {
-        const response = await fetch('https://ulqifq6ii2.execute-api.us-east-1.amazonaws.com/scx-dsp/dspresource')
-        const jsonData = await response.json()
-        const actualData = JSON.parse(jsonData.body);
-        setCourseData(actualData);
-        console.log(actualData)
-      } catch (error) {
-        // Handle error
-      }
-    };
+    // const getData = async () => {
+    //   try {
+    //     const response = await fetch('https://asxucwg1u7.execute-api.us-east-1.amazonaws.com/dev-test-1/dspfetchalldata')
+    //     const jsonData = await response.json()
+    //     const actualData = JSON.parse(jsonData.body);
+    //     setCourseData(actualData);
+    //     console.log(actualData)
+    //   } catch (error) {
+    //     // Handle error
+    //   }
+    // };
 
     getData();
   }, []);
@@ -44,11 +58,10 @@ const Courses = () => {
   const handleSaveData = async (data) => {
     try {
         // Make the POST request to save data to the backend
-        const response = await fetch('https://ulqifq6ii2.execute-api.us-east-1.amazonaws.com/scx-dsp/dspresource', {
+        const response = await fetch('https://zaoz973vnj.execute-api.us-east-1.amazonaws.com/scx-dsp/dspsavedata', {
             method: 'POST',
             headers: {
                 'Content-Type': 'application/json',
-                'Access-Control-Allow-Headers': '*',
             },
             body: JSON.stringify(data),
         });
@@ -61,6 +74,7 @@ const Courses = () => {
             // Handle error response
             console.error('Failed to save data:', response.status);
         }
+        getData();
     } catch (error) {
         console.error('Error saving data:', error);
     }
@@ -68,29 +82,39 @@ const Courses = () => {
 
 //to delete a single item
   const handleDeleteData = async (idToDelete) => {
-    // setCourseData((prevData) => prevData.filter((item) => item.id !== id));
-    try {
-      const response = await fetch('https://nhwx7j6qaa.execute-api.us-east-1.amazonaws.com/scx-dev-1/dspupdatedata/${idToDelete}', {
-        method: 'DELETE',
-        headers: {
-          'Content-Type': 'application/json',
-          // Include any additional headers if needed
-        },
-      });
+    const isConfirmed = window.confirm('Are you sure want to delete the item ?')
+    if (isConfirmed) {
+      try {
 
-      if (response.ok) {
-        // Successful deletion
-        setDeleted(true)
-        console.log('deleted the data')
-        // Update your state or fetch data again as needed
-      } else {
-        // Handle errors
-        console.error('Error deleting data:', response.statusText);
+        const response = await fetch(`https://nhwx7j6qaa.execute-api.us-east-1.amazonaws.com/scx-dev-1/dspupdatedata/${idToDelete}`, {
+          method: 'DELETE',
+          headers: {
+            'Content-Type': 'application/json',
+            // Include any additional headers if needed
+          },
+        });
+
+
+        if (response.ok) {
+          // Successful deletion
+          setDeleted(true)
+          toast.success('Item deleted successfully!', { position: toast.POSITION.TOP_CENTER });
+          console.log('deleted the data')
+          // Update your state or fetch data again as needed
+        } else {
+          // Handle errors
+          toast.warning('Failed to delete data, please try again!', { position: toast.POSITION.TOP_CENTER });
+          console.error('Error deleting data:', response.statusText);
+        }
+        getData();
+      } catch (error) {
+        toast.error('Error deleting data', { position: toast.POSITION.TOP_CENTER })
+        console.error('Error deleting data:', error.message);
       }
-    } catch (error) {
-      console.error('Error deleting data:', error.message);
+    } else {
+       toast.info('Deletion cancelled.', { position: toast.POSITION.TOP_CENTER })
     }
-  };
+  }
 
   //to delete all the data 
   const handleDeleteAll = async () => {
@@ -110,6 +134,7 @@ const Courses = () => {
             // Handle errors
             console.error('Error deleting all data:', response.statusText);
         }
+        getData();
     } catch (error) {
         console.error('Error deleting all data:', error.message);
     }
